Extract protocols into separate files
