+++ Sources/JWTKeychain/Support/RequestUpdateable.swift
@@ -1,0 +1,5 @@
+import HTTP
+
+public protocol RequestUpdateable {
+    func update(request: Request) throws
+}
