+++ Sources/JWTKeychain/Support/TokenGenerator.swift
@@ -1,0 +1,8 @@
+import Authentication
+import Fluent
+
+public protocol TokenGenerator {
+    func generateToken<E>(
+        for: E
+    ) throws -> Token where E: PasswordAuthenticatable, E: Entity
+}
